Consolidated option tag types in a single options header
