+++ src/resonanceReconstruction/rmatrix/CompoundSystem.hpp
@@ -15,11 +15,10 @@
 #include "range/v3/view/all.hpp"
 #include "range/v3/view/transform.hpp"
 #include "resonanceReconstruction/Quantity.hpp"
+#include "resonanceReconstruction/rmatrix/options.hpp"
 #include "resonanceReconstruction/rmatrix/Map.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionID.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionChannelID.hpp"
-#include "resonanceReconstruction/rmatrix/Formalism.hpp"
-#include "resonanceReconstruction/rmatrix/BoundaryOption.hpp"
 #include "resonanceReconstruction/rmatrix/SpinGroup.hpp"
 
 namespace njoy {
