--- src/resonanceReconstruction/rmatrix/Formalism.hpp
@@ -1,22 +1,0 @@
-#ifndef NJOY_R2_RMATRIX_FORMALISM
-#define NJOY_R2_RMATRIX_FORMALISM
-
-// system includes
-
-// other includes
-
-namespace njoy {
-namespace resonanceReconstruction {
-namespace rmatrix {
-
-  // R-matrix formalism options
-  struct SingleLevelBreitWigner {};
-  struct MultiLevelBreitWigner {};
-  struct ReichMoore {};
-  struct GeneralRMatrix {};
-
-} // rmatrix namespace
-} // resonanceReconstruction namespace
-} // njoy namespace
-
-#endif
