+++ src/resonanceReconstruction/rmatrix/SpinGroup.hpp
@@ -26,10 +26,9 @@
 #include "range/v3/view/zip_with.hpp"
 #include "range/v3/view/zip.hpp"
 #include "resonanceReconstruction/Quantity.hpp"
+#include "resonanceReconstruction/rmatrix/options.hpp"
 #include "resonanceReconstruction/rmatrix/Map.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionChannelID.hpp"
-#include "resonanceReconstruction/rmatrix/Formalism.hpp"
-#include "resonanceReconstruction/rmatrix/BoundaryOption.hpp"
 #include "resonanceReconstruction/rmatrix/ParticleChannel.hpp"
 #include "resonanceReconstruction/rmatrix/ParticleChannelData.hpp"
 #include "resonanceReconstruction/rmatrix/ResonanceTable.hpp"
