+++ src/resonanceReconstruction/rmatrix/calculateCoulombPhaseShift.hpp
@@ -7,7 +7,7 @@
 #include "range/v3/numeric/accumulate.hpp"
 #include "range/v3/view/indices.hpp"
 #include "range/v3/view/transform.hpp"
-#include "resonanceReconstruction/rmatrix/ChannelType.hpp"
+#include "resonanceReconstruction/rmatrix/options.hpp"
 
 namespace njoy {
 namespace resonanceReconstruction {
