+++ src/resonanceReconstruction/rmatrix/fromENDF.hpp
@@ -9,8 +9,7 @@
 #include "resonanceReconstruction/endf.hpp"
 #include "resonanceReconstruction/endf.hpp"
 #include "resonanceReconstruction/Quantity.hpp"
-#include "resonanceReconstruction/rmatrix/Formalism.hpp"
-#include "resonanceReconstruction/rmatrix/BoundaryOption.hpp"
+#include "resonanceReconstruction/rmatrix/options.hpp"
 #include "resonanceReconstruction/rmatrix/CompoundSystem.hpp"
 #include "resonanceReconstruction/rmatrix/legacy/resolved/CompoundSystem.hpp"
 #include "resonanceReconstruction/rmatrix/legacy/unresolved/CompoundSystem.hpp"
