+++ src/resonanceReconstruction/rmatrix/legacy/resolved/CompoundSystem.hpp
@@ -7,8 +7,8 @@
 #include "range/v3/range/conversion.hpp"
 #include "range/v3/action/sort.hpp"
 #include "range/v3/action/unique.hpp"
+#include "resonanceReconstruction/rmatrix/options.hpp"
 #include "resonanceReconstruction/rmatrix/Map.hpp"
-#include "resonanceReconstruction/rmatrix/Formalism.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionID.hpp"
 #include "resonanceReconstruction/rmatrix/legacy/CompoundSystemBase.hpp"
 #include "resonanceReconstruction/rmatrix/legacy/resolved/SpinGroup.hpp"
