+++ src/resonanceReconstruction/rmatrix/legacy/resolved/SpinGroup.hpp
@@ -13,8 +13,8 @@
 #include "range/v3/numeric/accumulate.hpp"
 #include "range/v3/view/zip_with.hpp"
 #include "resonanceReconstruction/Quantity.hpp"
+#include "resonanceReconstruction/rmatrix/options.hpp"
 #include "resonanceReconstruction/rmatrix/Map.hpp"
-#include "resonanceReconstruction/rmatrix/Formalism.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionID.hpp"
 #include "resonanceReconstruction/rmatrix/legacy/Data.hpp"
 #include "resonanceReconstruction/rmatrix/legacy/SpinGroupBase.hpp"
